requirements.txt
arcticdb==3.6.0
boto3==1.28.85
botocore==1.31.85
flask==3.0.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1
pydantic==2.5.2
pyyaml==6.0.1
requests==2.31.0
gunicorn==21.2.0
